+++ views/js/front/order-opc-phone-handler.js
@@ -9,21 +9,32 @@ $(document).ready(function (){
 
 function handlePhoneNumber(selector)
 {
+    if ($('.dpd-phone-block') !== undefined) {
+
     var phone = selector.find('input[name="dpd-phone"]').val();
     var phoneArea = selector.find('select[name="dpd-phone-area"] option:selected').val();
 
     var termsAndConditionsCheckbox =  $("#uniform-cgv input");
-    if (!$.isNumeric(phone)) {
+    if (!$.isNumeric(phone) || !phone) {
         $('.dpd-checkout-phone-container .error-message').removeClass('hidden');
+        selector.find('input[name="dpd-phone"]').css('border-color', 'red');
         termsAndConditionsCheckbox.attr("disabled", true);
         return false;
     } else {
         $('.dpd-checkout-phone-container .error-message').addClass('hidden');
         termsAndConditionsCheckbox.removeAttr("disabled");
+        selector.find('input[name="dpd-phone"]').css('border-color', 'initial');
     }
 
     saveSelectedPhoneNumber(phone, phoneArea)
+    }
 }
+
+$('#uniform-cgv').on('click', function (){
+    if ($('.dpd-phone-block') !== undefined) {
+        handlePhoneNumber($('.dpd-phone-block'));
+    }
+});
 
 function saveSelectedPhoneNumber(phoneNumber, phoneArea) {
     $.ajax(dpdHookAjaxUrl, {